bot/requirements.txt
# Telegram Bot Framework
aiogram==3.20.0.post0
magic-filter==1.0.12

# Async HTTP
aiohttp==3.11.18
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiosignal==1.3.2

# Web Framework
Flask==3.1.1
Quart==0.20.0
Werkzeug==3.1.3
Hypercorn==0.17.3
click==8.2.1
blinker==1.9.0
itsdangerous==2.2.0
Jinja2==3.1.6

# gRPC & Protocol Buffers
grpcio==1.62.0
grpcio-tools==1.62.0
protobuf==5.27.1

# Database
pymongo==4.13.0

# Configuration
python-dotenv==1.1.0

# HTTP Client
requests==2.32.3
urllib3==2.4.0

# Dependencies
typing-extensions==4.13.2
typing-inspection==0.4.1
pydantic==2.11.5
pydantic_core==2.33.2
MarkupSafe==3.0.2

# Network & Encoding
dnspython==2.7.0
charset-normalizer==3.4.2
certifi==2025.4.26
idna==3.10

# Utilities
attrs==25.3.0
colorama==0.4.6
frozenlist==1.6.0
multidict==6.4.4
yarl==1.20.0
propcache==0.3.1
priority==2.0.0

# HTTP/2 Support
h2==4.2.0
h11==0.16.0
hpack==4.1.0
hyperframe==6.1.0
wsproto==1.2.0

# Annotations
annotated-types==0.7.0
